DBOS "client context" (#53)
This PR adds workflow management features usable by a "client context",
that is, a DBOSContext object that is initialized but not launched. The
benefits of doing so is being able to interact with DBOS workflows
without having the process dequeue tasks, recover workflows, or run an
admin server.

Features added:
- Resume workflow
- Fork workflow
- Enqueue workflow by name
- Add ability to dismiss output/input from list workflows

Because the registry is indexed by workflow FQN, we need to keep another
map from workflow "custom" name to workflow FQN. This should have been
tested in PR that added custom workflow names.